Ensure that singleStepMoe is always reset to false after clicking greenFlag.
This fixes a bug where if you stopped a program where singleStepMode was
true (which is always true when Scratch is paused), then restarted the
program via clicking the greenFlag - Scratch would single step and
immediately pause. The correct behavior is to run normally until the
first breakpoint is hit. The cause of the error was singleStepMode =
true being carried over.